+++ app/models/raif/agent.rb
@@ -38,6 +38,7 @@ module Raif
     include Raif::Concerns::HasLlm
     include Raif::Concerns::HasRequestedLanguage
     include Raif::Concerns::HasAvailableModelTools
+    include Raif::Concerns::HasRuntimeDuration
     include Raif::Concerns::InvokesModelTools
     include Raif::Concerns::AgentInferenceStats
     include Raif::Concerns::RunWith